+++ packages/library-base/src/commonMain/kotlin/io/realm/kotlin/internal/RealmObjectCompanion.kt
@@ -23,11 +23,10 @@ import kotlin.reflect.KClass
 import kotlin.reflect.KMutableProperty1
 import kotlin.reflect.KProperty1
 
-// TODO MEDIATOR/API-INTERNAL Consider adding type parameter for the class
 // TODO Public due to being a transitive dependency to Mediator.
 @Suppress("VariableNaming")
-public interface RealmObjectCompanion<T : TypedRealmObject> {
-    public val `io_realm_kotlin_class`: KClass<T>
+public interface RealmObjectCompanion<out T : TypedRealmObject> {
+    public val `io_realm_kotlin_class`: KClass<out T>
     public val `io_realm_kotlin_className`: String
     public val `io_realm_kotlin_fields`: Map<String, KProperty1<BaseRealmObject, Any?>>
     public val `io_realm_kotlin_primaryKey`: KMutableProperty1<*, *>?